+++ packages/terra-form-select/src/search/Frame.jsx
@@ -232,10 +232,9 @@ class Frame extends React.Component {
     this.setState({ isFocused: false, focusedByTouch: false });
 
     if (this.state.hasSearchChanged) {
-      this.setState({
-        searchValue: this.props.display,
-        hasSearchChanged: false,
-      });
+      this.setState((prevState) => ({
+        searchValue: prevState.searchValue.trim(),
+      }));
     }
 
     this.closeDropdown();